refactor into util/git
