--- lib/config/simple-git.ts
@@ -1,10 +1,0 @@
-import { SimpleGitOptions } from 'simple-git';
-
-export function simpleGitConfig(): Partial<SimpleGitOptions> {
-  return {
-    completion: {
-      onClose: true,
-      onExit: false,
-    },
-  };
-}
