+++ lib/datasource/crate/index.ts
@@ -2,12 +2,12 @@ import hasha from 'hasha';
 import Git from 'simple-git';
 import { join } from 'upath';
 import { getGlobalConfig } from '../../config/global';
-import { simpleGitConfig } from '../../config/simple-git';
 import { logger } from '../../logger';
 import { ExternalHostError } from '../../types/errors/external-host-error';
 import * as memCache from '../../util/cache/memory';
 import * as packageCache from '../../util/cache/package';
 import { privateCacheDir, readFile } from '../../util/fs';
+import { simpleGitConfig } from '../../util/git/config';
 import { Http } from '../../util/http';
 import * as cargoVersioning from '../../versioning/cargo';
 import type { GetReleasesConfig, Release, ReleaseResult } from '../types';
