+++ lib/manager/git-submodules/extract.ts
@@ -2,9 +2,9 @@ import URL from 'url';
 import Git, { SimpleGit } from 'simple-git';
 import upath from 'upath';
 import { getGlobalConfig } from '../../config/global';
-import { simpleGitConfig } from '../../config/simple-git';
 import * as datasourceGitRefs from '../../datasource/git-refs';
 import { logger } from '../../logger';
+import { simpleGitConfig } from '../../util/git/config';
 import { getHttpUrl, getRemoteUrlWithToken } from '../../util/git/url';
 import type { ExtractConfig, PackageFile } from '../types';
 import { GitModule } from './types';
